# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DriverHistoryRead": {
        "description": "One month's km for a driver. Computed, never stored.",
        "properties": {
          "driver_id": {
            "format": "uuid",
            "title": "Driver Id",
            "type": "string"
          },
          "km": {
            "title": "Km",
            "type": "number"
          },
          "month": {
            "title": "Month",
            "type": "integer"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "driver_id",
          "year",
          "month",
          "km"
        ],
        "title": "DriverHistoryRead",
        "type": "object"
      },
...
      "DriverRead": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "address": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Address",
            "type": "string"
          },
          "auth_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Id"
          },
          "availability": {
            "items": {
              "type": "boolean"
            },
            "title": "Availability",
            "type": "array"
          },
          "car_make_model": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Car Make Model",
            "type": "string"
          },
          "driver_id": {
            "format": "uuid",
            "title": "Driver Id",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "readOnly": true,
            "title": "Full Name",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "license_plate": {
            "maxLength": 20,
            "minLength": 1,
            "title": "License Plate",
            "type": "string"
          },
          "note_chain_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Chain Id"
          },
          "partner_driver_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Driver Name"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Phone",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "license_plate",
          "car_make_model",
          "address",
          "driver_id",
          "user_id",
          "auth_id",
          "first_name",
          "last_name",
          "email",
          "role",
          "full_name"
        ],
        "title": "DriverRead",
        "type": "object"
      },
...
      "DriverUpdate": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "address": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "availability": {
            "anyOf": [
              {
                "items": {
                  "type": "boolean"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability"
          },
          "car_make_model": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Car Make Model"
          },
          "first_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "license_plate": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Plate"
          },
          "partner_driver_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Driver Name"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 20,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "title": "DriverUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API for the Food4Kids application",
    "title": "Food4Kids API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/drivers/": {
      "get": {
        "description": "Get all drivers, optionally filter by driver_id or email",
        "operationId": "get_drivers",
        "parameters": [
          {
            "description": "Filter by driver ID",
            "in": "query",
            "name": "driver_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by driver ID",
              "title": "Driver Id"
            }
          },
          {
            "description": "Filter by email",
            "in": "query",
            "name": "email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by email",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriverRead"
                  },
                  "title": "Response Get Drivers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Drivers",
        "tags": [
          "drivers"
        ]
      }
    },
...
    "/drivers/{driver_id}": {
      "delete": {
        "description": "Delete a driver by ID.\n\nTheir routes and mileage adjustments are detached (driver_id SET NULL),\nso the driver's km stop counting toward anyone.",
        "operationId": "delete_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Driver",
        "tags": [
          "drivers"
        ]
      },
      "get": {
        "description": "Get a single driver by ID",
        "operationId": "get_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Driver",
        "tags": [
          "drivers"
        ]
      },
      "put": {
        "description": "Update an existing driver",
        "operationId": "update_driver",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Driver",
        "tags": [
          "drivers"
        ]
      }
    },
    "/drivers/{driver_id}/history/": {
      "get": {
        "description": "Get monthly km totals, derived from the driver's frozen routes plus\nmanual adjustments (bucketed by drive_date month).\nRules:\n- No year, no month: return all months with activity\n- Year only: return all months for that year\n- Year + month: return specific month\n- Month without year: 400 error",
        "operationId": "get_driver_history",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2100,
                  "minimum": 2025,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 12,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriverHistoryRead"
                  },
                  "title": "Response Get Driver History",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Driver History",
        "tags": [
          "driver-history"
        ]
      }
    },
    "/drivers/{driver_id}/history/adjustments": {
      "get": {
        "description": "A driver's manual mileage adjustments, newest first (audit view).\nRoute-derived km isn't listed here — see the driver's routes for that.",
        "operationId": "get_driver_mileage_adjustments",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriverMileageAdjustmentRead"
                  },
                  "title": "Response Get Driver Mileage Adjustments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Driver Mileage Adjustments",
        "tags": [
          "driver-history"
        ]
      },
      "post": {
        "description": "Post a signed manual mileage adjustment (admin-only).\n\nkm is a delta (negative to remove over-credited distance) and a note\nexplaining the correction is required. Totals are derived from routes\nplus adjustments, so corrections compose with route credits.",
        "operationId": "create_driver_mileage_adjustment",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverMileageAdjustmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverMileageAdjustmentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Driver Mileage Adjustment",
        "tags": [
          "driver-history"
        ]
      }
    },
    "/drivers/{driver_id}/history/summary": {
      "get": {
        "description": "Get lifetime and current year KM summary for a driver",
        "operationId": "get_driver_history_summary",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Driver Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverHistorySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Driver History Summary",
        "tags": [
          "driver-history"
        ]
      }
    },
    "/drivers/{driver_id}/history/{year}/export": {
      "get": {
        "description": "Export history for all drivers for a specific year. Includes data from that year and the previous year.\n\n- driver_id: Must be \"all\"\n- year: The year to export data for",
        "operationId": "export_all_drivers_history",
        "parameters": [
          {
            "in": "path",
            "name": "driver_id",
            "required": true,
            "schema": {
              "title": "Driver Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export All Drivers History",
        "tags": [
          "driver-history"
        ]
      }
    },
...
    "/reports/monthly/{year}/{month}/totals": {
      "get": {
        "description": "Return total distance driven and total deliveries for the month.",
        "operationId": "get_monthly_totals",
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "month",
            "required": true,
            "schema": {
              "title": "Month",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyTotalsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Monthly Totals",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/routes": {
      "get": {
        "description": "Get routes with pagination and optional filtering for unassigned routes and date range.\nReturns routes with their drive dates - routes can appear multiple times for different dates.\nWhen unassigned_only is False, returns all routes (no assignment filter).\nWhen unassigned_only is True, returns only routes that are unassigned for the given route group.\n\nRequires a driver or admin caller.\nAdmins may scope to any driver via driver_id (or omit it for all routes).\nDrivers are always scoped to their own routes: omitting driver_id returns\ntheir own routes, and requesting another driver's is rejected.",
        "operationId": "get_routes",
        "parameters": [
          {
            "description": "If true, only return unassigned routes. If false, return all routes regardless of assignment status.",
            "in": "query",
            "name": "unassigned_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "If true, only return unassigned routes. If false, return all routes regardless of assignment status.",
              "title": "Unassigned Only",
              "type": "boolean"
            }
          },
          {
            "description": "Filter route groups from this date",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "description": "Filter route groups from this date",
              "title": "Start Date",
              "type": "string"
            }
          },
          {
            "description": "Filter route groups until this date",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "description": "Filter route groups until this date",
              "title": "End Date",
              "type": "string"
            }
          },
          {
            "description": "Order by drive_date: 'asc' (default, oldest-first) for the upcoming feed, 'desc' (most-recent-first) for the past feed.",
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "description": "Order by drive_date: 'asc' (default, oldest-first) for the upcoming feed, 'desc' (most-recent-first) for the past feed.",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "description": "Page number (1-indexed)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-indexed)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Number of items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Number of items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Filter routes by assigned driver. Admins may pass any driver_id, or omit it to get all routes. Drivers are always scoped to themselves: omitting it returns their own routes, and passing another driver's id is rejected with 403.",
            "in": "query",
            "name": "driver_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter routes by assigned driver. Admins may pass any driver_id, or omit it to get all routes. Drivers are always scoped to themselves: omitting it returns their own routes, and passing another driver's id is rejected with 403.",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_RouteWithDateRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Routes",
        "tags": [
          "routes"
        ]
      }
    },
...
  }
}